feat: update config utils with new config implementation - Botconfig now contains some getters and a method to add opt-out url - utils : contains generic function to resolve config elements
